+++ src/main/java/cz/ger/ffng/FFNGSurface.java
@@ -13,9 +13,9 @@ public class FFNGSurface {
         return BitmapFactory.decodeStream(FFNG.files.internal(file).read(), null, options);
     }
 
-    static Bitmap newSurface(FFNGFont font, String text, int frontColor, int bgColor, int outlineWidth) {
+    static Bitmap newSurface(FFNGFont font, String text, int frontColor, int bgColor, float outlineWidth) {
         Rect bounds = font.getBounds(text);
-        Bitmap bmp = Bitmap.createBitmap(bounds.width() + 2 * outlineWidth, bounds.height() + 2 * outlineWidth + 4, Bitmap.Config.ARGB_8888);
+        Bitmap bmp = Bitmap.createBitmap(bounds.width() + (int)(2.f * outlineWidth), bounds.height() + (int)(2.f * outlineWidth) + 4, Bitmap.Config.ARGB_8888);
         Canvas canvas = new Canvas(bmp);
         font.render(canvas, text, frontColor, bgColor, outlineWidth);
         return bmp;
