+++ src/main/jni/FFNGFont.h
@@ -15,7 +15,7 @@ public:
 	static int init();
 	static void quit();
 	void sizeUTF8(const char *text, int *width, int *height);
-	SDL_Surface* renderUTF8_Shaded(const char *text, Color fontColor, Color bgColor, int outlineWidth=0);
+	SDL_Surface* renderUTF8_Shaded(const char *text, Color fontColor, Color bgColor, float outlineWidth=0);
 private:
 	int getWidth(const char *text);
 	int getHeight(const char *text);
